Fix birthday, cursor pagination, and state saving
Make the `birthday` field under the `customers` stream accept strings that
do not format to datetimes (ie `0000-12-23T00:00:00Z`)

Fix querying with a cursor to include the previous query. We were
receiving `INVALID_REQUEST_ERROR` because we did not include the
start_date with the query.

Also moved the saving of the `start_date` bookmark to after we clear out
the `cursor` bookmark because we need the previously used `start_date` to
use the cursor.